move context updating to delib-context, inc round on done, ...otherProps not appropriate in steps
