+++ app/components/deliberation-context.js
@@ -1,13 +1,15 @@
-import React, { createContext, useCallback, useState, useRef } from 'react'
+import React, { createContext, useCallback, useState, useRef, useEffect } from 'react'
 import setOrDeleteByMutatePath from '../lib/set-or-delete-by-mutate-path'
+import socketApiSubscribe from '../socket-apis/socket-api-subscribe'
 
 export const DeliberationContext = createContext({})
 export default DeliberationContext
 
 export function DeliberationContextProvider(props) {
   const local = useRef({}).current // can't be in deriver becasue "Error: Rendered more hooks than during the previous render."
+  const { defaultValue = {} } = props
+  const { discussionId, userId } = defaultValue
   const [data, setData] = useState(() => {
-    const { defaultValue = {} } = props
     return deriveReducedPointList({ reducedPointList: [], ...defaultValue }, local)
   })
   const upsert = useCallback(
@@ -22,6 +24,32 @@ export function DeliberationContextProvider(props) {
     },
     [setData]
   )
+  useEffect(() => {
+    if (!discussionId) return
+    // steps are looking for userId in the context, if the user is not logged in to start, context needs to be updated
+    upsert({ discussionId, userId })
+
+    function onSubscribeHandler(data) {
+      if (data.uInfo) {
+        let round = data.uInfo.length - 1
+        data.round = round
+        if (data.uInfo[round].groupings?.length > 0) data.groupIdsLists = cloneDeep(data.uInfo[round].groupings)
+      }
+      upsert(data)
+    }
+
+    function onUpdateHandler(data) {
+      upsert(data)
+    }
+    socketApiSubscribe('subscribe-deliberation', discussionId, onSubscribeHandler, onUpdateHandler)
+    // after the socket is idle for too lone, the connection is closed to save server resources
+    // if the user start to interact again, the connection will reopen and we need to re subscribe
+    window.socket.on('connect', () => {
+      console.log('Reconnected to socket')
+      socketApiSubscribe('subscribe-deliberation', discussionId, onSubscribeHandler, onUpdateHandler)
+    })
+  }, [discussionId, upsert, userId])
+
   return <DeliberationContext.Provider value={{ data, upsert }}>{props.children}</DeliberationContext.Provider>
 }
